requirements.txt
streamlit
crewai
openai
firecrawl
python-dotenv
requests